freelancerTech 완성

+++ backend/src/main/java/com/back/domain/freelancer/freelancerTech/service/FreelancerTechService.java
@@ -1,8 +1,9 @@
 package com.back.domain.freelancer.freelancerTech.service;
 
 import com.back.domain.freelancer.freelancer.entity.Freelancer;
+import com.back.domain.freelancer.freelancer.service.FreelancerFinder;
 import com.back.domain.freelancer.freelancerTech.dto.FreelancerTechAddDto;
-import com.back.domain.freelancer.freelancerTech.dto.MyTechListResponseDto;
+import com.back.domain.freelancer.freelancerTech.dto.FreelancerTechDto;
 import com.back.domain.freelancer.freelancerTech.entity.FreelancerTech;
 import com.back.domain.freelancer.freelancerTech.repository.FreelancerTechRepository;
 import com.back.domain.tech.entity.Tech;
@@ -12,36 +13,59 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class FreelancerTechService {
     private final FreelancerTechRepository freelancerTechRepository;
     private final TechRepository techRepository;
-
-    @Transactional
-    public long addMyTech(Optional<Freelancer> freelancer, FreelancerTechAddDto dto) {
-        Tech tech = techRepository.findById(dto.id()).get();
-        FreelancerTech freelancerTech = new FreelancerTech(freelancer.get(), tech, dto.techLevel());
-        freelancerTechRepository.save(freelancerTech);
-        return freelancerTech.getId();
-    }
+    private final FreelancerFinder freelancerFinder;
 
     @Transactional(readOnly = true)
-    public List<MyTechListResponseDto> findTechsByFreelancer(Optional<Freelancer> freelancer) {
-        List<FreelancerTech> freelancerTechs = freelancerTechRepository.findFreelancerTechByFreelancer(freelancer);
+    public List<FreelancerTechDto> findTechsByFreelancerId(Long freelancerId) {
+        freelancerFinder.findFreelancerByMemberId(freelancerId);    // 프리랜서 존재 여부 확인
+
+        List<FreelancerTech> freelancerTechs = freelancerTechRepository.findByFreelancerId(freelancerId);
         return freelancerTechs.stream()
-                .map(MyTechListResponseDto::new)
+                .map(FreelancerTechDto::new)
                 .toList();
     }
 
     @Transactional
-    public long deleteTechById(long id) {
-        FreelancerTech tech = freelancerTechRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 기술입니다"));
+    public FreelancerTech addMyTech(Long memberId, FreelancerTechAddDto dto) {
+        Freelancer freelancer = freelancerFinder.findFreelancerByMemberId(memberId);
+        Tech tech = techRepository.findById(dto.techId()).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 기술입니다"));
+
+        if(freelancerTechRepository.existsByFreelancerAndTech(freelancer, tech)) {
+            throw new IllegalArgumentException("이미 추가된 기술입니다");
+        }
+
+        FreelancerTech freelancerTech = new FreelancerTech(freelancer, tech, dto.techLevel());
+
+        return freelancerTechRepository.save(freelancerTech);
+    }
+
+    @Transactional
+    public void update(Long id, String techLevel, Long memberId) {
+        FreelancerTech freelancerTech = freelancerTechRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 기술입니다"));
+
+        Freelancer freelancer = freelancerFinder.findFreelancerByMemberId(memberId);
+        if(!(freelancerTech.getFreelancer().getId() == freelancer.getId())) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
+
+        freelancerTech.update(techLevel);
+    }
+
+    @Transactional
+    public void delete(long id, Long memberId) {
+        FreelancerTech freelancerTech = freelancerTechRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 기술입니다"));
+
+        Freelancer freelancer = freelancerFinder.findFreelancerByMemberId(memberId);
+        if(!(freelancerTech.getFreelancer().getId() == freelancer.getId())) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
 
         freelancerTechRepository.deleteById(id);
-        return tech.getId();
     }
 }
